rename: search페이지 경로 수정
