+++ src/stories/SearchPage.stories.tsx
@@ -1,4 +1,4 @@
-import SearchPage from '@pages/SearchPage';
+import SearchPage from '@pages/searchPage/SearchPage';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
